+++ src/Plugin/Unipay/Shortcut/ScanShortcut.php
@@ -20,37 +20,37 @@ class ScanShortcut implements ShortcutInterface
      */
     public function getPlugins(array $params): array
     {
-        $typeMethod = Str::camel($params['_type'] ?? 'default').'Plugins';
+        $typeMethod = Str::camel($params['_action'] ?? 'default').'Plugins';
 
         if (method_exists($this, $typeMethod)) {
             return $this->{$typeMethod}();
         }
 
-        throw new InvalidParamsException(Exception::SHORTCUT_MULTI_TYPE_ERROR, "Scan type [{$typeMethod}] not supported");
+        throw new InvalidParamsException(Exception::SHORTCUT_MULTI_ACTION_ERROR, "Scan action [{$typeMethod}] not supported");
     }
 
-    public function defaultPlugins(): array
+    protected function defaultPlugins(): array
     {
         return [
             ScanNormalPlugin::class,
         ];
     }
 
-    public function preAuthPlugins(): array
+    protected function preAuthPlugins(): array
     {
         return [
             ScanPreAuthPlugin::class,
         ];
     }
 
-    public function preOrderPlugins(): array
+    protected function preOrderPlugins(): array
     {
         return [
             ScanPreOrderPlugin::class,
         ];
     }
 
-    public function feePlugins(): array
+    protected function feePlugins(): array
     {
         return [
             ScanFeePlugin::class,