+++ tests/Plugin/Unipay/Shortcut/CancelShortcutTest.php
@@ -32,15 +32,15 @@ class CancelShortcutTest extends TestCase
     {
         self::assertEquals([
             \Yansongda\Pay\Plugin\Unipay\QrCode\CancelPlugin::class,
-        ], $this->plugin->getPlugins(['_type' => 'qr_code']));
+        ], $this->plugin->getPlugins(['_action' => 'qr_code']));
     }
 
     public function testFoo()
     {
         self::expectException(InvalidParamsException::class);
-        self::expectExceptionCode(Exception::SHORTCUT_MULTI_TYPE_ERROR);
-        self::expectExceptionMessage('Cancel type [fooPlugins] not supported');
+        self::expectExceptionCode(Exception::SHORTCUT_MULTI_ACTION_ERROR);
+        self::expectExceptionMessage('Cancel action [fooPlugins] not supported');
 
-        $this->plugin->getPlugins(['_type' => 'foo']);
+        $this->plugin->getPlugins(['_action' => 'foo']);
     }
 }