+++ tests/TestCase.php
@@ -2,7 +2,7 @@
 
 namespace Yansongda\Pay\Tests;
 
-use DI\ContainerBuilder;
+use Hyperf\Pimple\ContainerFactory;
 use Hyperf\Utils\ApplicationContext;
 use Yansongda\Pay\Pay;
 
@@ -138,9 +138,9 @@ class TestCase extends \PHPUnit\Framework\TestCase
             ]
         ];
 
-        // 未在 hyperf 框架内，所以 sdk 没有 container, 手动设置一个
-        if (class_exists(ApplicationContext::class) && class_exists(ContainerBuilder::class)) {
-            ApplicationContext::setContainer((new ContainerBuilder())->build());
+        // hyperf 单测时，未在 hyperf 框架内，所以 sdk 没有 container, 手动设置一个
+        if (class_exists(ApplicationContext::class) && class_exists(ContainerFactory::class)) {
+            ApplicationContext::setContainer((new ContainerFactory())());
         }
 
         Pay::config($config);